remove django_snippets, move sessionIdleTimeout to distributed/middleware
Possible fix for #4713
